refactor presenters
