--- app/presenters/locale_command_handler_presenter.rb
@@ -1,11 +1,0 @@
-class LocaleCommandHandlerPresenter < BasePresenter
-  include MessageInterface
-
-  def formatted_text
-    I18n.t("telegram_webhooks.commands.set_locale.ask", locale:)
-  end
-
-  def inline_keyboard
-    Buttons::ForSetLocale.build
-  end
-end
