+++ app/presenters/media_generator/button_request_presenters/image_processed_message/for_prompt_to_image.rb
