+++ app/presenters/media_generator/button_request_presenters/image_processed_message/presenter_selector.rb
